app/service/project_info/requirements.txt
functions_framework>=3.8.2
google-cloud-resource-manager==1.14.1
google-cloud-bigquery>=3.30.0
google-cloud-logging>=3.11.4
google-cloud-monitoring>=2.27.0
protobuf>=5.29.3
#pytest>=8.3.5
#pytest-mock>=3.14.0
